Move enrichments to Response folder.
