+++ src/ApplicationCore/Entities/FraudProtectionApiModels/Response/Enrichments.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-namespace Contoso.FraudProtection.ApplicationCore.Entities.FraudProtectionApiModels
+namespace Contoso.FraudProtection.ApplicationCore.Entities.FraudProtectionApiModels.Response
 {
     public class Enrichments
     {
